+++ nsplusthon/client/soroushclient.py
@@ -1,12 +1,12 @@
 from . import (
-    AccountMethods, AuthMethods, DownloadMethods, DialogMethods, ChatMethods,
+    ContactMethods, AccountMethods, AuthMethods, DownloadMethods, DialogMethods, ChatMethods,
     BotMethods, MessageMethods, UploadMethods, ButtonMethods, UpdateMethods,
     MessageParseMethods, UserMethods, SoroushPlusBaseClient
 )
 
 
 class SoroushClient(
-    AccountMethods, AuthMethods, DownloadMethods, DialogMethods, ChatMethods,
+    ContactMethods, AccountMethods, AuthMethods, DownloadMethods, DialogMethods, ChatMethods,
     BotMethods, MessageMethods, UploadMethods, ButtonMethods, UpdateMethods,
     MessageParseMethods, UserMethods, SoroushPlusBaseClient
 ):